+++ src/main/java/com/vladoose/nir/service/MedEquipmentService.java
@@ -2,7 +2,9 @@ package com.vladoose.nir.service;
 
 import com.vladoose.nir.entity.MedEquipment;
 import com.vladoose.nir.entity.TenderLot;
+import com.vladoose.nir.exception.BadRequestException;
 import com.vladoose.nir.exception.NotFoundException;
+import com.vladoose.nir.repository.ApplyItemRepository;
 import com.vladoose.nir.repository.MedEquipmentRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,9 +15,11 @@ import java.util.List;
 public class MedEquipmentService {
 
     private final MedEquipmentRepository repository;
+    private final ApplyItemRepository applyItemRepository;
 
-    public MedEquipmentService(MedEquipmentRepository repository) {
+    public MedEquipmentService(MedEquipmentRepository repository, ApplyItemRepository applyItemRepository) {
         this.repository = repository;
+        this.applyItemRepository = applyItemRepository;
     }
 
     public List<MedEquipment> findAll() {
@@ -34,6 +38,9 @@ public class MedEquipmentService {
 
     @Transactional
     public void deleteById(Long id) {
+        if (applyItemRepository.existsByMedEquipmentId(id)) {
+            throw new BadRequestException("Невозможно удалить оборудование: оно используется в позициях заявок.");
+        }
         repository.deleteById(id);
     }
